+++ Deprecated/AppController+Debug.swift
@@ -10,27 +10,8 @@ import Foundation
 import UIKit
 
 extension AppController {
-
     func showDebug() {
         let controller = UINavigationController(rootViewController: DebugViewController())
         self.present(controller, animated: true, completion: nil)
     }
-
-#if DEBUG
-    // swiftlint:disable override_in_extension
-    override var canBecomeFirstResponder: Bool {
-        true
-    }
-
-    override func motionEnded(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
-        guard motion == .motionShake else { return }
-        self.presentDebugViewController()
-    }
-
-    func presentDebugViewController() {
-        guard self.presentingViewController == nil else { return }
-        self.showDebug()
-    }
-    // swiftlint:enable override_in_extension
-#endif
 }