+++ Source/Controller/DirectoryViewController.swift
@@ -12,7 +12,7 @@ import Analytics
 import CrashReporting
 import SwiftUI
 
-class DirectoryViewController: ContentViewController, AboutTableViewDelegate, HelpDrawerHost {
+class DirectoryViewController: ContentViewController, AboutTableViewDelegate, HelpDrawerViewControllerHost {
     
     /// A model for the various table view sections
     enum Section: Int, CaseIterable {
@@ -345,12 +345,16 @@ extension DirectoryViewController: UITableViewDelegate {
         switch section {
         case .communityPubs:
             let star = communityPubs[indexPath.row]
-            let view = IdentityView(identity: star.feed).environmentObject(BotRepository.shared)
+            let view = IdentityView(identity: star.feed)
+                .environmentObject(BotRepository.shared)
+                .environmentObject(AppController.shared)
             let controller = UIHostingController(rootView: view)
             self.navigationController?.pushViewController(controller, animated: true)
         case .users:
             let identity = searchFilter
-            let view = IdentityView(identity: identity).environmentObject(BotRepository.shared)
+            let view = IdentityView(identity: identity)
+                .environmentObject(BotRepository.shared)
+                .environmentObject(AppController.shared)
             let controller = UIHostingController(rootView: view)
             self.navigationController?.pushViewController(controller, animated: true)
         case .posts:
@@ -364,7 +368,9 @@ extension DirectoryViewController: UITableViewDelegate {
             )
         case .network:
             let identity = self.people[indexPath.row].identity
-            let view = IdentityView(identity: identity).environmentObject(BotRepository.shared)
+            let view = IdentityView(identity: identity)
+                .environmentObject(BotRepository.shared)
+                .environmentObject(AppController.shared)
             let controller = UIHostingController(rootView: view)
             self.navigationController?.pushViewController(controller, animated: true)
         }