+++ Source/UI/About/IdentityView.swift
@@ -17,6 +17,9 @@ struct IdentityView: View {
     @EnvironmentObject
     private var botRepository: BotRepository
 
+    @EnvironmentObject
+    private var appController: AppController
+
     @State
     private var about: About?
     
@@ -55,7 +58,7 @@ struct IdentityView: View {
             errorMessage != nil
         } set: { _ in
             errorMessage = nil
-            AppController.shared.dismiss(animated: true)
+            appController.dismiss(animated: true)
         }
     }
 
@@ -347,5 +350,6 @@ struct IdentityView_Previews: PreviewProvider {
             .preferredColorScheme(.dark)
         }
         .environmentObject(BotRepository.fake)
+        .environmentObject(AppController.shared)
     }
 }