+++ Source/UI/About/MessageButtonStyle.swift
@@ -1,0 +1,72 @@
+//
+//  MessageButtonStyle.swift
+//  Planetary
+//
+//  Created by Martin Dutra on 9/12/22.
+//  Copyright © 2022 Verse Communications Inc. All rights reserved.
+//
+
+import SwiftUI
+
+/// Use this button style when using a MessageView as a Button label
+/// It will give it the shadows and a pressed state
+struct MessageButtonStyle: ButtonStyle {
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .offset(y: configuration.isPressed ? 3 : 0)
+            .compositingGroup()
+            .shadow(color: .cardBorderBottom, radius: 0, x: 0, y: 4)
+            .shadow(
+                color: .cardShadowBottom,
+                radius: configuration.isPressed ? 5 : 10,
+                x: 0,
+                y: configuration.isPressed ? 1 : 4
+            )
+    }
+}
+
+struct MessageButtonStyle_Previews: PreviewProvider {
+    static var messageValue: MessageValue {
+        MessageValue(
+            author: "@QW5uYVZlcnNlWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFg=.ed25519",
+            content: Content(
+                from: Post(
+                    blobs: nil,
+                    branches: nil,
+                    hashtags: nil,
+                    mentions: nil,
+                    root: nil,
+                    text: .loremIpsum(words: 10)
+                )
+            ),
+            hash: "",
+            previous: nil,
+            sequence: 0,
+            signature: .null,
+            claimedTimestamp: 0
+        )
+    }
+
+    static var message: Message {
+        var message = Message(
+            key: "@unset",
+            value: messageValue,
+            timestamp: 0
+        )
+        message.metadata = Message.Metadata(
+            author: Message.Metadata.Author(about: About(about: .null, name: "Mario")),
+            replies: Message.Metadata.Replies(count: 0, abouts: Set()),
+            isPrivate: false
+        )
+        return message
+    }
+
+    static var previews: some View {
+        Button {
+            print("Clicked")
+        } label: {
+            MessageView(message: message)
+        }
+        .buttonStyle(MessageButtonStyle())
+    }
+}