+++ CHANGELOG.md
@@ -6,6 +6,14 @@ We define "Noteworthy changes" as 1) user-facing features or bugfixes 2) signifi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.12] 2022-12-13
+
+- Fixed a bug that caused hashtags to not display any posts. #1021
+
+## [1.3.11] 2022-12-09
+
+- Added new redesigned Home screen. #974
+- Added new redesigned Hashtag screen. #975
 ## Unreleased
 - Display aliases on profile screen #979
 ## Unreleased